requirements.txt
Pillow
keyboard
numpy
pyautogui
pygetwindow
easyocr
pywin32